requirements.txt
Flask
goose-extractor
